Fix dialogue trigger not properly triggering again for the same object
Clears the _previousHits HashSet on the exit command for triggers, properly allowing the same object to be triggered again once the object is not in contact with it anymore.
DialogueTriggers now force the object to have a Collider and that the Collider is set to trigger. If it is not a trigger, it will force it to be a trigger and log a warning on the console.